backend/requirements.txt
Flask
pymongo[srv]
flask-cors
python-dotenv